Restructure Xcode project by functional area instead of object type.
