+++ WeatherApi/WeatherApi/Current/Wind/WindModel.swift
