+++ WeatherApi/WeatherApi/Current/Wind/WindViewModel.swift
