+++ WeatherApi/WeatherApi/DataSources/ApiErrorType.swift
