+++ WeatherApi/WeatherApi/Localization/Assets.xcassets/AppIcon.appiconset/Contents.json
