+++ WeatherApi/WeatherApi/Localization/Assets.xcassets/Contents.json
