+++ WeatherApi/WeatherApi/Localization/Assets.xcassets/cellBackground.colorset/Contents.json
