+++ WeatherApi/WeatherApi/Localization/Assets.xcassets/high.colorset/Contents.json
