+++ WeatherApi/WeatherApi/Localization/Assets.xcassets/low.colorset/Contents.json
