+++ WeatherApi/WeatherApi/Localization/Localizable.xcstrings
