+++ WeatherApi/WeatherApi/Settings/AppSettings.swift
